pyproject.toml
[project]
name = "hackrx-hackathon"
version = "0.1.0"
description = "HackRx 6.0 - RAG-based Q&A API for document processing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "chromadb>=0.4.18",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "pypdf2>=3.0.1",
]
